+++ packages/video_player/video_player/CHANGELOG.md
@@ -1,5 +1,6 @@
 ## NEXT
 
+* Automatic Bitrate Streaming support to Android and iOS.
 * Updates minimum supported SDK version to Flutter 3.32/Dart 3.8.
 * Updates README to reflect currently supported OS versions for the latest
   versions of the endorsed platform implementations.